מ

+++ src/main/java/com/ashcollege/repository/UserRepository.java
@@ -1,12 +1,11 @@
 package com.ashcollege.repository;
 
 import com.ashcollege.entities.UserEntity;
-import com.ashcollege.entities.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
 
     boolean existsByMail(String email);
